feat(factorial-number): Add factorial operation of a number. Update all values to support 64 bit length.

+++ CalculationModel.go
@@ -1,11 +1,11 @@
 package main
 
 type CalcNumbers struct {
-	Number1 float32
-	Number2 float32
+	Number1 float64
+	Number2 float64
 }
 
 type CalcWholeNumbers struct {
-	Number1 int32
-	Number2 int32
+	Number1 int64
+	Number2 int64
 }
